pyproject.toml
[project]
name = "booking-hotels"
version = "0.1.0"
description = ""
authors = [
    {name = "Semen Sobolev",email = "sema.bx@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "ipython (>=8.32.0,<9.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
